+++ backend/app-orchestrator-service/src/main/kotlin/rpc/SyncthingController.kt
@@ -1,6 +1,5 @@
 package dk.sdu.cloud.app.orchestrator.rpc
 
-import dk.sdu.cloud.accounting.util.asController
 import dk.sdu.cloud.app.orchestrator.api.*
 import dk.sdu.cloud.app.orchestrator.services.*
 import dk.sdu.cloud.calls.server.RpcServer