+++ frontend-web/k8.kts
@@ -3,7 +3,7 @@ package dk.sdu.cloud.k8
 
 bundle {
     name = "webclient"
-    version = "2022.3.25"
+    version = "2023.1.0"
 
     withAmbassador(null) {
         addSimpleMapping("/api/auth-callback")