+++ frontend-web/webclient/app/Project/ContextSwitcher.tsx
@@ -97,7 +97,7 @@ function _ContextSwitcher(props: ContextSwitcherReduxProps & DispatchProps): JSX
                     )}
                     {props.activeProject || response.data.items.length > 0 ? <Divider /> : null}
                     <Text onClick={() => navigate("/projects")}>Manage projects</Text>
-                    <Text onClick={() => navigate("/projects")}>
+                    <Text onClick={() => navigate(`/projects/${projectId ?? "My Workspace"}`)}>
                         {projectId ? "Manage active project" : "Manage my workspace"}
                     </Text>
                 </BoxForPadding>
@@ -139,6 +139,10 @@ function onProjectUpdated(navigate: NavigateFunction, runThisFunction: () => voi
     } else if (splitPath.length === 3) {
         if (splitPath[0] === "app" && splitPath[1] === "projects") {
             navigate(`/projects/${projectId}`);
+        }
+    } else if (splitPath.length === 5) {
+        if (splitPath[0] === "app" && splitPath[2] === "grants" && splitPath[3] == "view") {
+            navigate(`/project/grants/ingoing/${projectId}`);
         }
     }
     initializeResources();